+++ coding/PonSat/main/src/libraries/logger/logger.h
@@ -10,7 +10,7 @@ class SerialLogger
 {
     public:
         void init();
-        void writeCanSat(String teamID, long time, double alt, uint32_t a, bool stp, bool spp, bool rcp, bool lnp);
+        void writeCanSat(String teamID, uint32_t time, uint16_t alt, uint16_t a, bool stp, bool spp, bool rcp, bool lnp);
 };
 
 class SDLogger
@@ -19,7 +19,7 @@ class SDLogger
     public:
         void init();
         void write(String str);
-        void writeCanSat(String teamID, long time, double alt, uint32_t ax, uint32_t ay, uint32_t az, uint32_t gx, uint32_t gy, uint32_t gz, uint32_t mx, uint32_t my, uint32_t mz, uint32_t pressure, uint32_t temp, bool stp, bool spp, bool rcp, bool lnp);
+        void writeCanSat(String teamID, uint32_t time, uint16_t alt, int16_t ax, int16_t ay, int16_t az, int16_t gx, int16_t gy, int16_t gz, int16_t mx, int16_t my, int16_t mz, uint16_t pressure, uint32_t temp, bool stp, bool spp, bool rcp, bool lnp);
 };
 
 #endif
